+++ src/AvePoint.Elements/Private/Wait-AvptShellContinue.ps1
@@ -1,0 +1,9 @@
+function Wait-AvptShellContinue {
+    [CmdletBinding()]
+    param(
+        [string] $Message = 'Press Enter to return to the dashboard'
+    )
+
+    Write-Host ''
+    [void] (Read-Host $Message)
+}